Make the taskkill run before the update step.
Small fix in try_job_dir.
Review URL: http://codereview.chromium.org/9390

git-svn-id: svn://svn.chromium.org/chrome/trunk@4878 0039d316-1c4b-4281-b951-d872f2087c98

+++ tools/buildbot/config/master.tryserver/try_job_dir.py
@@ -101,7 +101,8 @@ class TryJobDir(Try_Jobdir):
       # that somebody is holding open. This was causing a Permission
       # Denied error on bear's win32-twisted1.3 buildslave.
       # Also, remove the destination file first.
-      os.remove(file_to)
+      if os.path.exists(file_to):
+        os.remove(file_to)
       os.rename(file_from, file_to)
       f = open(file_to, "r")
